third_party/WebKit/LayoutTests/fast/ime/inputmethodcontext-composition-crash-expected.txt

This tests a case that a Composition object outlives InputMethodContext.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS composition.selectionEnd is 0
PASS successfullyParsed is true

TEST COMPLETE
